File moved to Logging -> Log .hpp/cpp
