+++ src/logonserver/LogonCommServer/LogonCommServer.cpp
@@ -21,7 +21,7 @@
 #include "LogonCommDefines.h"
 
 #include <Common.hpp>
-#include <Log.hpp>
+#include <Logging/Log.hpp>
 #include <Logging/Logger.hpp>
 #include <Realm/RealmManager.hpp>
 #include <Server/Master.hpp>
