+++ src/shared/Logging/Log.cpp
@@ -3,14 +3,14 @@ Copyright (c) 2014-2023 AscEmu Team <http://www.ascemu.org>
 This file is released under the MIT license. See README-MIT for more information.
 */
 
-#include "Log.hpp"
-#include "Logging/Logger.hpp"
-#include "Utilities/Util.hpp"
-#include "Config/Config.h"
-
 #include <iostream>
 #include <cstdarg>
 #include <string>
+
+#include "Log.hpp"
+#include "Logger.hpp"
+#include "Utilities/Util.hpp"
+#include "Config/Config.h"
 
 //////////////////////////////////////////////////////////////////////////////////////////
 // World functions
