+++ src/shared/Logging/Log.hpp
