+++ src/world/Management/GameEventMgr.cpp
@@ -5,7 +5,7 @@ This file is released under the MIT license. See README-MIT for more information
 
 
 #include "GameEventMgr.h"
-#include "Log.hpp"
+#include "Logging/Log.hpp"
 #include "Server/World.h"
 #include "Server/MainServerDefines.h"
 #include "GameEvent.h"
