+++ src/world/Server/Packets/Handlers/QueryHandler.cpp
@@ -9,7 +9,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include "Server/Packets/SmsgNameQueryResponse.h"
 #include "Server/Packets/SmsgGameobjectQueryResponse.h"
 #include "Server/Packets/SmsgQueryTimeResponse.h"
-#include "Log.hpp"
+#include "Logging/Log.hpp"
 #include "Management/ObjectMgr.h"
 #include "Storage/MySQLDataStore.hpp"
 #include "Server/Packets/CmsgCreatureQuery.h"
