+++ src/world/pchWorld.hpp
@@ -18,7 +18,7 @@ This file is released under the MIT license. See README-MIT for more information
 #include <Common.hpp>
 #include <Config/Config.h>
 #include <WorldPacket.h>
-#include <Log.hpp>
+#include <Logging/Log.hpp>
 #include <Logging/Logger.hpp>
 #include <ByteBuffer.h>
 #include <Config/Config.h>
